+++ admin/controller/extension/payment/mollie/base.php
@@ -86,8 +86,40 @@ class ControllerExtensionPaymentMollieBase extends Controller
 				DB_PREFIX
 			));
 
+		$this->db->query("ALTER TABLE `" . DB_PREFIX . "order` MODIFY `payment_method` VARCHAR(255) NOT NULL;");
+
 		// Just install all modules while we're at it.
 		$this->installAllModules();
+		$this->cleanUp();
+	}
+
+	/**
+	 * Clean up files that are not needed for the running version of OC.
+	 */
+	public function cleanUp()
+	{
+		$adminThemeDir = DIR_APPLICATION . 'view/template/extension/payment/';
+		$catalogThemeDir = DIR_CATALOG . 'view/theme/default/template/extension/payment/';
+
+		// Remove old template from previous version.
+		unlink($adminThemeDir . 'mollie_2.tpl');
+
+		if (MollieHelper::isOpenCart3x()) {
+			unlink($adminThemeDir . 'mollie_1.tpl');
+			unlink($adminThemeDir . 'mollie.tpl');
+			unlink($catalogThemeDir . 'mollie_return.tpl');
+			unlink($catalogThemeDir . 'mollie_checkout_form.tpl');
+		} elseif (MollieHelper::isOpenCart2x()) {
+			unlink($adminThemeDir . 'mollie_1.tpl');
+			unlink($adminThemeDir . 'mollie.twig');
+			unlink($catalogThemeDir . 'mollie_return.twig');
+			unlink($catalogThemeDir . 'mollie_checkout_form.twig');
+		} else {
+			unlink($adminThemeDir . 'mollie.tpl');
+			unlink($adminThemeDir . 'mollie.twig');
+			unlink($catalogThemeDir . 'mollie_return.twig');
+			unlink($catalogThemeDir . 'mollie_checkout_form.twig');
+		}
 	}
 
 	/**
@@ -97,9 +129,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 	{
 		// Load models.
 		$extensions = $this->getExtensionModel();
-		$this->load->model("user/user_group");
-
-		$user_id = $this->user->getGroupId();
+		$user_id = $this->getUserId();
 
 		foreach (MollieHelper::$MODULE_NAMES as $module_name)
 		{
@@ -137,18 +167,6 @@ class ControllerExtensionPaymentMollieBase extends Controller
 	}
 
 	/**
-	 * Get the extension installation handler.
-	 *
-	 * @return Model
-	 */
-	protected function getExtensionModel ()
-	{
-		$this->load->model("setting/extension");
-
-		return $this->model_setting_extension;
-	}
-
-	/**
 	 * Render the payment method's settings page.
 	 */
 	public function index ()
@@ -163,6 +181,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		$this->document->setTitle($this->language->get("heading_title"));
 
 		$shops = $this->getMultiStores();
+		$code = MollieHelper::getModuleCode();
 		$this->retrieveMultiStoreConfigs();
 
 		// Call validate method on POST
@@ -177,12 +196,12 @@ class ControllerExtensionPaymentMollieBase extends Controller
 				
 				foreach (MollieHelper::$MODULE_NAMES as $module_name)
 				{
-					$status = "payment_mollie_" . $module_name . "_status";
+					$status = $code . "_" . $module_name . "_status";
 
 					$post[$status] = (isset($post[$status]) && $post[$status] == "on") ? 1 : 0;
 				}
 
-				$this->model_setting_setting->editSetting("payment_mollie", $post, $store['id']);
+				$this->model_setting_setting->editSetting($code, $post, $store['id']);
 
 				// Migrate old settings if needed. We used to use "ideal" as setting group, but Opencart 2 requires us to use "mollie".
 				$this->model_setting_setting->deleteSetting("ideal", $store['id']);
@@ -194,11 +213,11 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		if ($doRedirect)
 		{
 			$this->session->data['success'] = $this->language->get("text_success");
-			$this->redirect($this->url->link("marketplace/extension", "type=payment&user_token=" . $this->session->data['user_token'], "SSL"));
+			$this->redirect($this->url->link($this->getExtensionsUri(), "type=payment&" . $this->getTokenUriPart(), "SSL"));
 		}
 
 		// Set data for template
-		$data['api_check_url']          = $this->url->link("extension/payment/mollie_" . static::MODULE_NAME . '/validate_api_key', "user_token=" . $this->session->data['user_token'], "SSL");
+		$data['api_check_url']          = $this->url->link("extension/payment/mollie_" . static::MODULE_NAME . '/validate_api_key', $this->getTokenUriPart(), "SSL");
 		$data['heading_title']          = $this->language->get("heading_title");
 		$data['title_global_options']   = $this->language->get("title_global_options");
 		$data['title_payment_status']   = $this->language->get("title_payment_status");
@@ -220,7 +239,6 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		$data['entry_description']              = $this->language->get("entry_description");
 		$data['entry_show_icons']               = $this->language->get("entry_show_icons");
 		$data['entry_show_order_canceled_page'] = $this->language->get("entry_show_order_canceled_page");
-		$data['entry_status']                   = $this->language->get("entry_status");
 		$data['entry_mod_status']               = $this->language->get("entry_mod_status");
 		$data['entry_comm_status']              = $this->language->get("entry_comm_status");
 		$data['entry_geo_zone']                 = $this->language->get("entry_geo_zone");
@@ -244,7 +262,6 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		$data['entry_activate']         = $this->language->get("entry_activate");
 		$data['entry_sort_order']       = $this->language->get("entry_sort_order");
 		$data['entry_support']          = $this->language->get("entry_support");
-		$data['entry_mstatus']          = $this->checkModuleStatus();
 		$data['entry_module']           = $this->language->get("entry_module");
 		$data['entry_version']          = $this->language->get("entry_version") . " " . MollieHelper::PLUGIN_VERSION;
 
@@ -254,6 +271,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		$data['tab_general']            = $this->language->get("tab_general");
 
 		$data['shops'] = $shops;
+		$data['code'] = $code;
 
 		// If there are errors, show the error.
 		if (isset($this->error['warning'])) {
@@ -266,7 +284,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 
 		foreach($shops as $store)
 		{
-			$data['stores'][$store['id']]['entry_cstatus'] = $this->checkCommunicationStatus(isset($store['payment_mollie_api_key']) ? $store['payment_mollie_api_key'] : null);
+			$data['stores'][$store['id']]['entry_cstatus'] = $this->checkCommunicationStatus(isset($store[$code . '_api_key']) ? $store[$code . '_api_key'] : null);
 
 			if (isset($this->error[$store['id']]['api_key'])) {
 				$data['stores'][$store['id']]['error_api_key'] = $this->error[$store['id']]['api_key'];
@@ -305,38 +323,38 @@ class ControllerExtensionPaymentMollieBase extends Controller
 		$data['breadcrumbs'] = array();
 
 		$data['breadcrumbs'][] = array(
-			"href"      => $this->url->link("common/dashboard", "user_token=" . $this->session->data['user_token'], "SSL"),
+			"href"      => $this->url->link("common/dashboard", $this->getTokenUriPart(), "SSL"),
 			"text"      => $this->language->get("text_home"),
 			"separator" => FALSE,
 		);
 
 		$data['breadcrumbs'][] = array(
-			"href"      => $this->url->link("marketplace/extension", "type=payment&user_token=" . $this->session->data['user_token'], "SSL"),
+			"href"      => $this->url->link($this->getExtensionsUri(), "type=payment&" . $this->getTokenUriPart(), "SSL"),
 			"text"      => $this->language->get("text_payment"),
 			"separator" => ' :: ',
 		);
 
 		$data['breadcrumbs'][] = array(
-			"href"      => $this->url->link("extension/payment/mollie_" . static::MODULE_NAME, "user_token=" . $this->session->data['user_token'], "SSL"),
+			"href"      => $this->url->link("extension/payment/mollie_" . static::MODULE_NAME, $this->getTokenUriPart(), "SSL"),
 			"text"      => $this->language->get("heading_title"),
 			"separator" => " :: ",
 		);
 
 		// Form action url
-		$data['action'] = $this->url->link("extension/payment/mollie_" . static::MODULE_NAME, "user_token=" . $this->session->data['user_token'], "SSL");
-		$data['cancel'] = $this->url->link("marketplace/extension", "type=payment&user_token=" . $this->session->data['user_token'], "SSL");
+		$data['action'] = $this->url->link("extension/payment/mollie_" . static::MODULE_NAME, $this->getTokenUriPart(), "SSL");
+		$data['cancel'] = $this->url->link($this->getExtensionsUri(), "type=payment&" . $this->getTokenUriPart(), "SSL");
 
 		// Load global settings. Some are prefixed with mollie_ideal_ for legacy reasons.
 		$settings = array(
-			"payment_mollie_api_key"                    => NULL,
-			"payment_mollie_ideal_description"          => "Order %",
-			"payment_mollie_show_icons"                 => FALSE,
-			"payment_mollie_show_order_canceled_page"   => TRUE,
-			"payment_mollie_ideal_pending_status_id"    => 1,
-			"payment_mollie_ideal_processing_status_id" => 2,
-			"payment_mollie_ideal_canceled_status_id"   => 7,
-			"payment_mollie_ideal_failed_status_id"     => 10,
-			"payment_mollie_ideal_expired_status_id"    => 14,
+			$code . "_api_key"                    => NULL,
+			$code . "_ideal_description"          => "Order %",
+			$code . "_show_icons"                 => FALSE,
+			$code . "_show_order_canceled_page"   => TRUE,
+			$code . "_ideal_pending_status_id"    => 1,
+			$code . "_ideal_processing_status_id" => 2,
+			$code . "_ideal_canceled_status_id"   => 7,
+			$code . "_ideal_failed_status_id"     => 10,
+			$code . "_ideal_expired_status_id"    => 14,
 		);
 
 		foreach($shops as $store)
@@ -381,7 +399,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 			catch (Mollie_API_Exception $e)
 			{
 				// If we have an unauthorized request, our API key is likely invalid.
-				if ($data['stores'][$store['id']]['payment_mollie_api_key'] !== NULL && strpos($e->getMessage(), "Unauthorized request") >= 0)
+				if ($data['stores'][$store['id']][$code . '_api_key'] !== NULL && strpos($e->getMessage(), "Unauthorized request") >= 0)
 				{
 					$data['stores'][$store['id']]['error_api_key'] = $this->language->get("error_api_key_invalid");
 				}
@@ -389,7 +407,6 @@ class ControllerExtensionPaymentMollieBase extends Controller
 
 			$data['stores'][$store['id']]['payment_methods'] = array();
 
-		
 			foreach (MollieHelper::$MODULE_NAMES as $module_name)
 			{
 				$payment_method = array();
@@ -399,31 +416,31 @@ class ControllerExtensionPaymentMollieBase extends Controller
 				$payment_method['allowed'] = in_array($module_name, $allowed_methods);
 
 				// Load module specific settings.
-				if (isset($this->request->post['stores'][$store['id']]['payment_mollie_' . $module_name . '_status']))
+				if (isset($this->request->post['stores'][$store['id']][$code . '_' . $module_name . '_status']))
 				{
-					$payment_method['status'] = ($this->request->post['stores'][$store['id']]['payment_mollie_' . $module_name . '_status'] == "on");
+					$payment_method['status'] = ($this->request->post['stores'][$store['id']][$code . '_' . $module_name . '_status'] == "on");
 				}
 				else
 				{
-					$payment_method['status'] = (bool) isset($this->data['stores'][$store['id']]["payment_mollie_" . $module_name . "_status"]) ? $this->data['stores'][$store['id']]["payment_mollie_" . $module_name . "_status"] : null;
+					$payment_method['status'] = (bool) isset($this->data['stores'][$store['id']][$code . "_" . $module_name . "_status"]) ? $this->data['stores'][$store['id']][$code . "_" . $module_name . "_status"] : null;
 				}
 
-				if (isset($this->request->post['stores'][$store['id']]['payment_mollie_' . $module_name . '_sort_order']))
+				if (isset($this->request->post['stores'][$store['id']][$code . '_' . $module_name . '_sort_order']))
 				{
-					$payment_method['sort_order'] = $this->request->post['stores'][$store['id']]['payment_mollie_' . $module_name . '_sort_order'];
+					$payment_method['sort_order'] = $this->request->post['stores'][$store['id']][$code . '_' . $module_name . '_sort_order'];
 				}
 				else
 				{
-					$payment_method['sort_order'] = isset($this->data['stores'][$store['id']]["payment_mollie_" . $module_name . "_sort_order"]) ? $this->data['stores'][$store['id']]["payment_mollie_" . $module_name . "_sort_order"] : null;
+					$payment_method['sort_order'] = isset($this->data['stores'][$store['id']][$code . "_" . $module_name . "_sort_order"]) ? $this->data['stores'][$store['id']][$code . "_" . $module_name . "_sort_order"] : null;
 				}
 
-				if (isset($this->request->post['stores'][$store['id']]['payment_mollie_' . $module_name . '_geo_zone']))
+				if (isset($this->request->post['stores'][$store['id']][$code . '_' . $module_name . '_geo_zone']))
 				{
-					$payment_method['geo_zone'] = $this->request->post['stores'][$store['id']]['payment_mollie_' . $module_name . '_geo_zone'];
+					$payment_method['geo_zone'] = $this->request->post['stores'][$store['id']][$code . '_' . $module_name . '_geo_zone'];
 				}
 				else
 				{
-					$payment_method['geo_zone'] = isset($this->data['stores'][$store['id']]["payment_mollie_" . $module_name . "_geo_zone"]) ? $this->data['stores'][$store['id']]["payment_mollie_" . $module_name . "_geo_zone"] : null;
+					$payment_method['geo_zone'] = isset($this->data['stores'][$store['id']][$code . "_" . $module_name . "_geo_zone"]) ? $this->data['stores'][$store['id']][$code . "_" . $module_name . "_geo_zone"] : null;
 				}
 
 				$data['stores'][$store['id']]['payment_methods'][$module_name] = $payment_method;
@@ -431,7 +448,12 @@ class ControllerExtensionPaymentMollieBase extends Controller
 			
 		}
 
-		$this->renderTemplate("extension/payment/mollie", $data, array(
+		$template = 'extension/payment/mollie';
+		if (!MollieHelper::isOpenCart2x()) {
+			$template = 'extension/payment/mollie_1';
+		}
+
+		$this->renderTemplate($template, $data, array(
 			"header",
 			"column_left",
 			"footer",
@@ -473,7 +495,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 					. '<br/><br/>'
 					. 'Please check the following conditions. You can ask your hosting provider to help with this.'
 					. '<ul>'
-					. '<li>Make sure outside connections to ' . ($client ? htmlspecialchars($client->getApiEndpoint()) : 'Mollie') . ' are not blocked.</li>'
+					. '<li>Make sure outside connections to ' . (isset($client) ? htmlspecialchars($client->getApiEndpoint()) : 'Mollie') . ' are not blocked.</li>'
 					. '<li>Make sure SSL v3 is disabled on your server. Mollie does not support SSL v3.</li>'
 					. '<li>Make sure your server is up-to-date and the latest security patches have been installed.</li>'
 					. '</ul><br/>'
@@ -498,67 +520,17 @@ class ControllerExtensionPaymentMollieBase extends Controller
 			$this->error['warning'] = $this->language->get("error_permission");
 		}
 
-		if (!$this->request->post['stores'][$store]['payment_mollie_api_key'])
+		if (!$this->request->post['stores'][$store][MollieHelper::getModuleCode() . '_api_key'])
 		{
 			$this->error[$store]['api_key'] = $this->language->get("error_api_key");
 		}
 
-		if (!$this->request->post['stores'][$store]['payment_mollie_ideal_description'])
+		if (!$this->request->post['stores'][$store][MollieHelper::getModuleCode() . '_ideal_description'])
 		{
 			$this->error[$store]['description'] = $this->language->get("error_description");
 		}
 		
 		return (count($this->error) == 0);
-	}
-
-	protected function checkModuleStatus ()
-	{
-		$need_files = array();
-
-		$mod_files  = array(
-			DIR_APPLICATION . "controller/extension/payment/mollie/base.php",
-			DIR_APPLICATION . "language/en-gb/extension/payment/mollie.php",
-			DIR_TEMPLATE . "extension/payment/mollie.twig",
-			DIR_CATALOG . "controller/extension/payment/mollie-api-client/",
-			DIR_CATALOG . "controller/extension/payment/mollie/base.php",
-			DIR_CATALOG . "language/en-gb/extension/payment/mollie.php",
-			DIR_CATALOG . "model/extension/payment/mollie/base.php",
-			DIR_CATALOG . "view/theme/default/template/extension/payment/mollie_checkout_form.twig",
-			DIR_CATALOG . "view/theme/default/template/extension/payment/mollie_return.twig",
-		);
-
-		foreach (MollieHelper::$MODULE_NAMES as $module_name)
-		{
-			$mod_files[] = DIR_APPLICATION . "controller/extension/payment/mollie_" . $module_name . ".php";
-			$mod_files[] = DIR_APPLICATION . "language/en-gb/extension/payment/mollie_" . $module_name . ".php";
-			$mod_files[] = DIR_CATALOG . "controller/extension/payment/mollie_" . $module_name . ".php";
-			$mod_files[] = DIR_CATALOG . "model/extension/payment/mollie_" . $module_name . ".php";
-		}
-
-		foreach ($mod_files as $file)
-		{
-			$realpath = realpath($file);
-
-			if (!file_exists($realpath))
-			{
-				$need_files[] = '<span class="text-danger">' . $file . '</span>';
-			}
-		}
-
-		if (!MollieHelper::apiClientFound())
-		{
-			$need_files[] = '<span class="text-danger">'
-				. 'API client not found. Please make sure you have installed the module correctly. Use the download '
-				. 'button on the <a href="https://github.com/mollie/OpenCart/releases/latest" target="_blank">release page</a>'
-				. '</span>';
-		}
-
-		if (count($need_files) > 0)
-		{
-			return $need_files;
-		}
-
-		return '<span class="text-success">OK</span>';
 	}
 
 	/**
@@ -610,15 +582,36 @@ class ControllerExtensionPaymentMollieBase extends Controller
 	 */
 	protected function renderTemplate ($template, $data, $common_children = array(), $echo = TRUE)
 	{
-		foreach ($common_children as $child)
-		{
-			$data[$child] = $this->load->controller("common/" . $child);
+		if (!MollieHelper::isOpenCart3x()) {
+			$template .= '.tpl';
 		}
 
-		$html = $this->load->view($template, $data);
+		if (MollieHelper::isOpenCart2x()) {
+			foreach ($common_children as $child) {
+				$data[$child] = $this->load->controller("common/" . $child);
+			}
 
-		if ($echo)
-		{
+			$html = $this->load->view($template, $data);
+		} else {
+			$this->template = $template;
+			$this->children = array();
+
+			foreach ($data as $field => $value) {
+				$this->data[$field] = $value;
+			}
+
+			foreach($common_children as $child) {
+				if ($child === 'column_left') {
+					continue;
+				}
+
+				$this->children[] = "common/" . $child;
+			}
+
+			$html = $this->render();
+		}
+
+		if ($echo) {
 			return $this->response->setOutput($html);
 		}
 
@@ -631,7 +624,7 @@ class ControllerExtensionPaymentMollieBase extends Controller
 	 */
 	protected function redirect ($url, $status = 302)
 	{
-		$this->response->redirect($url, $status);
+		$this->response->redirect(str_replace('&amp;', '&', $url), $status);
 	}
 
 	/**
@@ -671,6 +664,65 @@ class ControllerExtensionPaymentMollieBase extends Controller
 			}
 			$this->data['stores'][$store['id']] = $newArrray;
 		}
-		
+	}
+
+	/**
+	 * Get the extension installation handler.
+	 *
+	 * @return Model
+	 */
+	protected function getExtensionModel()
+	{
+		if (MollieHelper::isOpenCart3x()) {
+			$this->load->model('setting/extension');
+			return $this->model_setting_extension;
+		}
+
+		if (MollieHelper::isOpenCart2x()) {
+			$this->load->model('extension/extension');
+			return $this->model_extension_extension;
+		}
+
+		$this->load->model('setting/extension');
+		return $this->model_setting_extension;
+	}
+
+	/**
+	 * @return string
+	 */
+	private function getExtensionsUri()
+	{
+		if (MollieHelper::isOpenCart3x()) {
+			return 'marketplace/extension';
+		}
+
+		if (MollieHelper::isOpenCart23x()) {
+			return 'extension/extension';
+		}
+
+		return 'extension/payment';
+	}
+
+	/**
+	 * @return string
+	 */
+	private function getTokenUriPart()
+	{
+		if (MollieHelper::isOpenCart3x()) {
+			return 'user_token=' . $this->session->data['user_token'];
+		}
+
+		return 'token=' . $this->session->data['token'];
+	}
+
+	private function getUserId()
+	{
+		$this->load->model('user/user_group');
+
+		if (method_exists($this->user, 'getGroupId')) {
+			return $this->user->getGroupId();
+		}
+
+		return $this->user->getId();
 	}
 }